requirements.txt
streamlit>=1.20.0
openai>=0.27.0
pandas>=1.3.0
matplotlib>=3.4.0
python-dotenv>=0.21.0
